Fixed bugs
Added memsim.py - per the instructions, I think thats what the TAs will
look for. Made some changes to it too, such that the simulation will
end if the time goes past MAXTIME.
Also fixed some crashing issues if no args are passed to the program,
as well as crashing if the user presses enter without entering any data
when prompted.

Maybe we need a readme?

Also, it appears that my IDE has added a bunch of cruft along with the
actual code. Plz ignore!